improvement:create modal code refactored
